Seed data progress. Cleaned up seeding files/folders.
