+++ controller/auth.controller.js
@@ -9,7 +9,7 @@
 const authUtil = require("../util/auth.util");
 const enumUtil = require("../util/enum.util");
 const jwt = require('jsonwebtoken');
-const { db } = require("../config/db");
+const { db } = require("../db/models/db");
 //Load required db models for querying
 const { Organizer, Attendee } = db.models;
