+++ controller/user.controller.js
@@ -7,7 +7,7 @@
 
 //Import dependencies
 const enumUtil = require("../util/enum.util");
-const { db } = require("../config/db");
+const { db } = require("../db/models/db");
 //Load required db models for querying
 const {Organizer, Attendee} = db.models;
