+++ db/models/event/event-image.model.js
@@ -7,6 +7,11 @@ const { DataTypes } = require('sequelize');
 
 module.exports = (sequelize) => {
   sequelize.define('EventImage', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
     url: {
         type: DataTypes.STRING,
         allowNull: false,
