+++ db/models/event/event.model.js
@@ -7,6 +7,11 @@ const { DataTypes } = require('sequelize');
 
 module.exports = (sequelize) => {
   sequelize.define('Event', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
     title: {
       type: DataTypes.STRING,
       allowNull: false
