+++ db/models/event/performer.model.js
@@ -7,6 +7,11 @@ const { DataTypes } = require('sequelize');
 
 module.exports = (sequelize) => {
   sequelize.define('Performer', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
     firstName: {
         type: DataTypes.STRING,
         allowNull: false
