+++ db/models/event/tag.model.js
@@ -1,20 +1,20 @@
 /**
- * Act model
-*/
+ * Tag model
+ */
 
 const { DataTypes } = require('sequelize');
 
 
 module.exports = (sequelize) => {
-  sequelize.define('Act', {
+  sequelize.define('Tag', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
     name: {
         type: DataTypes.STRING,
         allowNull: false
-      },
-      dateFormed: {
-        type: DataTypes.DATEONLY,
-        allowNull:true
-      },
-
+      }
   });
 };
