+++ db/models/event/ticket-type.model.js
@@ -7,6 +7,11 @@ const { DataTypes } = require('sequelize');
 
 module.exports = (sequelize) => {
   sequelize.define('TicketType', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
     name: {
         type: DataTypes.STRING,
         allowNull: false
