+++ db/models/user/attendee.model.js
@@ -3,10 +3,15 @@
  */
 
   const { DataTypes } = require('sequelize');
-  const AuthUtil = require("../../util/auth.util");
+  const AuthUtil = require("../../../util/auth.util");
 
   module.exports = (sequelize) => {
     sequelize.define('Attendee', {
+      id: {
+        type: DataTypes.UUID,
+        defaultValue: DataTypes.UUIDV4,
+        primaryKey:true,
+      },
         firstName: {
             type: DataTypes.STRING,
             allowNull: false
