+++ db/models/user/organizer.model.js
@@ -3,11 +3,16 @@
  */
 
   const { DataTypes } = require('sequelize');
-  const AuthUtil = require("../../util/auth.util");
+  const AuthUtil = require("../../../util/auth.util");
 
 
 module.exports = (sequelize) => {
   sequelize.define('Organizer', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey:true,
+    },
       firstName: {
         type: DataTypes.STRING,
         allowNull: false
