--- models/event/tag.model.js
@@ -1,15 +1,0 @@
-/**
- * Tag model
- */
-
-const { DataTypes } = require('sequelize');
-
-
-module.exports = (sequelize) => {
-  sequelize.define('Tag', {
-    name: {
-        type: DataTypes.STRING,
-        allowNull: false
-      }
-  });
-};
