+++ package.json
@@ -16,6 +16,7 @@
     "cookie": "^0.5.0",
     "cookie-signature": "^1.0.6",
     "cors": "^2.8.5",
+    "cross-env": "^7.0.3",
     "crypto": "^1.0.1",
     "debug": "^2.6.9",
     "depd": "^2.0.0",
@@ -74,8 +75,8 @@
     "vary": "^1.1.2"
   },
   "scripts": {
-    "dev": "nodemon app.js",
-    "prod": "node app.js"
+    "dev": "cross-env NODE_ENV=development nodemon app.js",
+    "prod": "cross-env NODE_ENV=production node app.js"
   },
   "author": "",
   "license": "ISC",
